+++ packages/f-services/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.7.0
+------------------------------
+*December 1, 2020*
+
+### Added
+- `es-ES` validations to postcode and phone validations.
+### Changed
+- `isValidPostcode` to accept locale.
+
+
 v1.6.0
 ------------------------------
 *December 1, 2020*